video-translator-service/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
google-cloud-storage==2.14.0
google-cloud-speech==2.26.0
google-cloud-translate==3.15.1
google-genai==0.3.0
demucs==4.0.1
torch --index-url https://download.pytorch.org/whl/cpu
torchaudio --index-url https://download.pytorch.org/whl/cpu
torchcodec